LayoutTests/fast/custom-elements/Document-defineElement.html

<!DOCTYPE html>
<html>
<head>
<title>Custom Elements: Extensions to Document interface</title>
<meta name="author" title="Ryosuke Niwa" href="mailto:rniwa@webkit.org">
<meta name="assert" content="document.defineElement should define a custom element">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<link rel='stylesheet' href='../../resources/testharness.css'>
</head>
<body>
<div id="log"></div>
<script>

test(function () {
    assert_true('defineElement' in Document.prototype, '"defineElement" exists on Document.prototype');
    assert_true('defineElement' in document, '"defineElement" exists on document');
}, 'Check the existence of defineElement on Document interface');

test(function () {
    class MyCustomElement extends HTMLElement {};

    assert_throws({'name': 'SyntaxError'}, function () { document.defineElement(null, MyCustomElement); },
        'document.defineElement must throw a SyntaxError if the tag name is null');
    assert_throws({'name': 'SyntaxError'}, function () { document.defineElement('', MyCustomElement); },
        'document.defineElement must throw a SyntaxError if the tag name is empty');
    assert_throws({'name': 'SyntaxError'}, function () { document.defineElement('abc', MyCustomElement); },
        'document.defineElement must throw a SyntaxError if the tag name does not contain "-"');
    assert_throws({'name': 'SyntaxError'}, function () { document.defineElement('a-Bc', MyCustomElement); },
        'document.defineElement must throw a SyntaxError if the tag name contains an upper case letter');

    var builtinTagNames = [
        'annotation-xml',
        'color-profile',
        'font-face',
        'font-face-src',
        'font-face-uri',
        'font-face-format',
        'font-face-name',
        'missing-glyph'
    ];

    for (var tagName of builtinTagNames) {
        assert_throws({'name': 'SyntaxError'}, function () { document.defineElement(tagName, MyCustomElement); },
            'document.defineElement must throw a SyntaxError if the tag name is "' + tagName + '"');
    }

}, 'document.defineElement should throw with an invalid name');

test(function () {
    class SomeCustomElement extends HTMLElement {};
    class OtherCustomElement extends HTMLElement {};

    document.defineElement('some-custom-element', SomeCustomElement);
    assert_throws({'name': 'NotSupportedError'}, function () { document.defineElement('some-custom-element', OtherCustomElement); },
        'document.defineElement must throw a NotSupportedError if the specified tag name is already used');

}, 'document.defineElement should throw when there is already a custom element of the same name');

test(function () {
    class AnotherCustomElement extends HTMLElement {};

    document.defineElement('another-custom-element', AnotherCustomElement);
    assert_throws({'name': 'NotSupportedError'}, function () { document.defineElement('some-other-element', AnotherCustomElement); },
        'document.defineElement must throw a NotSupportedError if the specified class already defines an element');

}, 'document.defineElement should throw when there is already a custom element with the same class');

test(function () {
    class SomeCustomElement extends HTMLElement {};

    var templateContentOwnerDocument = document.createElement('template').content.ownerDocument;
    assert_throws({'name': 'NotSupportedError'}, function () {
        templateContentOwnerDocument.defineElement('some-custom-element', SomeCustomElement);
    });

}, 'document.defineElement must throw a NotSupportedError when the context object is an associated inert template document');

test(function () {
    class SomeCustomElement extends HTMLElement {};

    var windowlessDocument = document.implementation.createHTMLDocument();
    assert_throws({'name': 'NotSupportedError'}, function () {
        windowlessDocument.defineElement('some-custom-element', SomeCustomElement);
    });

}, 'document.defineElement must throw a NotSupportedError when the context object is created by DOMImplementation.createHTMLDocument');

test(function () {
    class SomeCustomElement extends HTMLElement {};

    var windowlessDocument = document.implementation.createDocument('http://www.w3.org/1999/xhtml', 'html', null)
    assert_throws({'name': 'NotSupportedError'}, function () {
        windowlessDocument.defineElement('some-custom-element', SomeCustomElement);
    });

}, 'document.defineElement must throw a NotSupportedError when the context object is created by DOMImplementation.createDocument');

test(function () {
    assert_throws({'name': 'TypeError'}, function () { document.defineElement('invalid-element', 1); },
        'document.defineElement must throw a TypeError when the element interface is a number');
    assert_throws({'name': 'TypeError'}, function () { document.defineElement('invalid-element', '123'); },
        'document.defineElement must throw a TypeError when the element interface is a string');
    assert_throws({'name': 'TypeError'}, function () { document.defineElement('invalid-element', {}); },
        'document.defineElement must throw a TypeError when the element interface is an object');
    assert_throws({'name': 'TypeError'}, function () { document.defineElement('invalid-element', []); },
        'document.defineElement must throw a TypeError when the element interface is an array');
}, 'document.defineElement should throw when the element interface is not a constructor');

test(function () {
    class MyCustomElement extends HTMLElement {};
    document.defineElement('my-custom-element', MyCustomElement);

    var instance = new MyCustomElement;
    assert_true(instance instanceof MyCustomElement,
        'An instance of a custom HTML element be an instance of the associated interface');

    assert_true(instance instanceof HTMLElement,
        'An instance of a custom HTML element must inherit from HTMLElement');

    assert_equals(instance.localName, 'my-custom-element',
        'An instance of a custom element must use the associated tag name');

    assert_equals(instance.namespaceURI, 'http://www.w3.org/1999/xhtml',
        'A custom element HTML must use HTML namespace');

}, 'document.defineElement should define an instantiatable custom element');

</script>
</body>
</html>
